+++ MiDAS/CHANGELOG.md
@@ -3,6 +3,10 @@
   previous top-associated allele as covariate, until there’s no more significant 
   alleles 
 
+## 0.0.0.9003 - 05/02/2019
++ G groups alleles are now accepted as properly formatted
++ G groups alleles can be now reduced, with warning
+
 ## 0.0.0.9002 - 27/02/2019
 + introduce analyzeHlaAssociations for performing statistical analysys of HLA
   alleles associations